+++ datacontract/export/rdf_converter.py
@@ -1,7 +1,8 @@
 from pydantic import BaseModel
 from rdflib import Graph, Literal, BNode, RDF, URIRef, Namespace
 
-from datacontract.model.data_contract_specification import DataContractSpecification
+from datacontract.model.data_contract_specification import \
+    DataContractSpecification
 
 
 def is_literal(property_name):
@@ -140,7 +141,7 @@ def add_info(contract, info, graph, dc, dcx):
     graph.add((bnode_info, dc.version, Literal(info.version)))
 
     # add owner
-    owner = URIRef(info.owner)
+    owner = Literal(info.owner)
     graph.add((bnode_info, dc.owner, owner))
 
     # add contact