+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added support for `delta` tables on S3 (#24)
 - Added new command `datacontract catalog` that generates a data contract catalog with an `index.html` file.
 
+### Fixed
+- RDF Export: Fix error if owner is not a URI/URN
+
+
 ## [0.10.1] - 2024-04-19
 
 ### Fixed